requirements.txt
pyyaml
dohq-artifactory
